### Added - visualization_aids::visualize_correlations function that plot heatmap of features' correlations. - documentation for the new methods. ### Changed - requirements dependencies. - visualization_aids::generate_decision_paths moved to xai::generate_decision_paths. - visualization_aids::visualize_features changed parameter name frame to data. ### Fixed - documentation fixes. - minor changes
