+++ .version
@@ -1,1 +1,1 @@
-version==1.5
+version==1.6
