+++ data-science-utils/meta.yaml
@@ -16,7 +16,7 @@ build:
 requirements:
   host:
     - joblib >=0.12
-    - matplotlib >2.02,<=3.0.3
+    - matplotlib >2.02
     - numpy >=1.3
     - pandas >=0.23.0
     - pip
@@ -26,7 +26,7 @@ requirements:
     - seaborn >=0.8.0
   run:
     - joblib >=0.12
-    - matplotlib >2.02,<=3.0.3
+    - matplotlib >2.02
     - numpy >=1.3
     - pandas >=0.23.0
     - pydotplus >=2.0.2
