+++ requirements-dev.txt
@@ -1,4 +1,4 @@
-pytest==5.3.2
+pytest==5.4.2
 pytest-cov==2.8.1
 nose==1.3.7
-coverage==4.5.4
+coverage==5.0
