+++ requirements-docs.txt
@@ -1,2 +1,2 @@
-sphinx==2.3.0
+sphinx==3.0.4
 sphinx_rtd_theme==0.4.3
