+++ requirements.txt
@@ -1,6 +1,6 @@
 numpy>=1.3
 pandas>=0.23.0
-matplotlib>2.02,<=3.0.3
+matplotlib>2.02
 seaborn>=0.8.0
 scikit-learn>=0.19.0
 pydotplus>=2.0.2
